+++ src/Exception/ErrorCode.php
@@ -18,6 +18,7 @@ use MyCLabs\Enum\Enum;
  * @method static ErrorCode REVIEWS_ARE_DISABLED()
  * @method static ErrorCode SENDER_IS_ALSO_RECIPIENT()
  * @method static ErrorCode COMPANY_HAS_NO_ADMINISTRATOR()
+ * @method static ErrorCode COMPANY_NOT_FOUND()
  */
 final class ErrorCode extends Enum
 {
@@ -28,4 +29,5 @@ final class ErrorCode extends Enum
     private const REVIEWS_ARE_DISABLED = 5;
     private const SENDER_IS_ALSO_RECIPIENT = 6;
     private const COMPANY_HAS_NO_ADMINISTRATOR = 7;
+    private const COMPANY_NOT_FOUND = 8;
 }