+++ tests/Discussion/DiscussionServiceTest.php
@@ -14,6 +14,7 @@ use Wizaplace\SDK\Discussion\Discussion;
 use Wizaplace\SDK\Discussion\DiscussionService;
 use Wizaplace\SDK\Discussion\Message;
 use Wizaplace\SDK\Exception\CompanyHasNoAdministrator;
+use Wizaplace\SDK\Exception\CompanyNotFound;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\ProductNotFound;
 use Wizaplace\SDK\Exception\SenderIsAlsoRecipient;
@@ -104,10 +105,10 @@ final class DiscussionServiceTest extends ApiTestCase
         $this->buildDiscussionService()->startDiscussionWithVendor(1);
     }
 
-    public function testStartDiscussionOnInexistantVendor()
+    public function testStartDiscussionOnInexistentVendor()
     {
         $this->expectExceptionCode(404);
-        $this->expectExceptionMessage('Company 404 has not been found.');
+        $this->expectException(CompanyNotFound::class);
 
         $this->discussionService->startDiscussionWithVendor(404);
     }